(PC-29252) refactor(share): Move types to appropriate file
