+++ src/features/auth/pages/signup/AccountCreated/AccountCreated.native.test.tsx
@@ -5,7 +5,7 @@ import { useAuthContext } from 'features/auth/context/AuthContext'
 import { navigateToHomeConfig } from 'features/navigation/helpers/navigateToHome'
 import { navigateFromRef } from 'features/navigation/navigationRef'
 import { ShareAppWrapper } from 'features/share/context/ShareAppWrapper'
-import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
+import { ShareAppModalType } from 'features/share/types'
 import { analytics } from 'libs/analytics'
 import { BatchUser } from 'libs/react-native-batch'
 import { render, fireEvent, waitFor, screen } from 'tests/utils'
