+++ src/features/identityCheck/pages/confirmation/BeneficiaryAccountCreated.native.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { BeneficiaryAccountCreated } from 'features/identityCheck/pages/confirmation/BeneficiaryAccountCreated'
 import { ShareAppWrapper } from 'features/share/context/ShareAppWrapper'
-import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
+import { ShareAppModalType } from 'features/share/types'
 import { BatchUser } from 'libs/react-native-batch'
 import { fireEvent, render, screen } from 'tests/utils'
