+++ src/features/internal/cheatcodes/pages/NavigationShareApp/NavigationShareApp.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native'
 
 import { Row } from 'features/internal/cheatcodes/components/Row'
 import { useShareAppContext } from 'features/share/context/ShareAppWrapper'
-import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
+import { ShareAppModalType } from 'features/share/types'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { PageHeaderSecondary } from 'ui/components/headers/PageHeaderSecondary'
