+++ src/features/share/context/ShareAppWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { memo, useCallback, useContext, useMemo, useState } from 'react'
 
-import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
 import { ShareAppModal } from 'features/share/pages/ShareAppModal'
+import { ShareAppModalType } from 'features/share/types'
 import { analytics } from 'libs/analytics'
 import { useModal } from 'ui/components/modals/useModal'
