+++ src/features/share/pages/ShareAppModal.tsx
@@ -3,10 +3,8 @@ import styled from 'styled-components/native'
 
 import { SHARE_APP_IMAGE_SOURCE } from 'features/share/components/shareAppImage'
 import { shareApp } from 'features/share/helpers/shareApp'
-import {
-  shareAppModalInformations,
-  ShareAppModalType,
-} from 'features/share/helpers/shareAppModalInformations'
+import { shareAppModalInformations } from 'features/share/helpers/shareAppModalInformations'
+import { ShareAppModalType } from 'features/share/types'
 import { analytics } from 'libs/analytics'
 import { useRemoteConfigContext } from 'libs/firebase/remoteConfig'
 import { ButtonQuaternaryBlack } from 'ui/components/buttons/ButtonQuaternaryBlack'
