+++ src/features/share/types.ts
@@ -6,3 +6,11 @@ export interface WebShareModalProps {
   shareContent: ShareContent
   dismissModal: () => void
 }
+
+export enum ShareAppModalType {
+  NOT_ELIGIBLE = 'NOT_ELIGIBLE',
+  BENEFICIARY = 'BENEFICIARY',
+  ON_BOOKING_SUCCESS = 'ON_BOOKING_SUCCESS',
+}
+
+export type ShareAppWordingVersion = 'default' | 'statistics' | 'short'
